% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ.R
\name{summ}
\alias{summ}
\title{Summation}
\usage{
summ(i, e)
}
\arguments{
\item{i}{the index expressed as "i:n" i=initial index, n= final index.}

\item{e}{The formula of the terms. The formula is expressed inside quotations and the only variable that can be used is i.}
}
\value{
the sum of terms in the sequence.
}
\description{
summ can calculate the sum of terms of a sequence (Description)
}
\details{
Summation notation is used in statistics to calculate a large sum of terms. Each consecutive number in the index is substituted as "i" in the formula and all the results are added up.(Details)
}
\examples{
summ(1:8,'2^i/i^2 + 3^i/i^3 + 4^i/i^4')
}
\author{
Kaily Rosal
}
