% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varpop.R
\name{varpop}
\alias{varpop}
\title{Population variation}
\usage{
varpop(x)
}
\arguments{
\item{x}{Your data (as a set of integers)}
}
\value{
Population variation
}
\description{
varpop can calculate the variation of a population (Description)
}
\details{
The variation of a population tells us how spread out the data is. It gives us how close the data points are to the mean. If the variation is small, they are close to the mean. If the variation is large, they are spread out from the mean. The population variance is useful when you do have all the available data. (Details)
}
\examples{
varpop(1:7)
}
\author{
Kaily Rosal
}
