% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varsam.R
\name{varsam}
\alias{varsam}
\title{Sample variation}
\usage{
varsam(x)
}
\arguments{
\item{x}{Your data (as a set of integers)}
}
\value{
Sample variation
}
\description{
varsam can calculate the variation of a sample (Description)
}
\details{
The variation of a sample tells us how spread out the data is. It gives us how close the data points are to the mean. If the variation is small, they are close to the mean. If the variation is large, they are spread out from the mean. The sample variance is useful when you do not have all the available data. (Details)
}
\examples{
varsam(1:7)
}
\author{
Kaily Rosal
}
