% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x.R
\docType{data}
\name{x}
\alias{x}
\title{Random set of integers.}
\format{
A set of 15 random integers picked from 1 to 15:
}
\usage{
x
}
\description{
An example of a set of integers to use for the functions in this package.
}
\keyword{datasets}
